Initial commit
This certainly contains a number of unused libraries.
This software is designed to run on a Teensy 4.1 controller with an MCUFriend KBV touchscreen. It can of course be ported over to any other touch screen, or to an Arduino Mega if necessary, but a traditional Uno won't have enough IO for this project.